+++ openeo-geotrellis/src/main/scala/org/openeo/geotrellis/GeneralUtils.scala
@@ -59,30 +59,30 @@ object GeneralUtils {
   }
 
   def cellTypeUnionWithNoData(leftCellType:CellType, rightCellType:CellType):CellType = {
-    def getNodataAndMax(cellType:CellType):(Option[Double],Double) = {
+    def getNodataMaxMin(cellType:CellType):(Option[Double],Double,Double) = {
       cellType match {
-        case BitCellType => (None, 1)
-        case ByteCellType => (None, Byte.MaxValue)
-        case UByteCellType => (None, 255)
-        case ShortCellType => (None, Short.MaxValue)
-        case UShortCellType => (None, 65535)
-        case IntCellType => (None, Int.MaxValue)
-        case FloatCellType => (None, Float.MaxValue)
-        case DoubleCellType => (None, Double.MaxValue)
-        case ByteConstantNoDataCellType => (Some(byteNODATA), Byte.MaxValue)
-        case UByteConstantNoDataCellType => (Some(ubyteNODATA), 255)
-        case ShortConstantNoDataCellType => (Some(shortNODATA), Short.MaxValue)
-        case UShortConstantNoDataCellType => (Some(ushortNODATA), 65535)
-        case IntConstantNoDataCellType => (Some(NODATA), Int.MaxValue)
-        case FloatConstantNoDataCellType => (Some(floatNODATA), Float.MaxValue)
-        case DoubleConstantNoDataCellType => (Some(doubleNODATA), Double.MaxValue)
-        case ct: ByteUserDefinedNoDataCellType => (Some(ct.noDataValue), Byte.MaxValue)
-        case ct: UByteUserDefinedNoDataCellType => (Some(ct.widenedNoData.asInt), 255)
-        case ct: ShortUserDefinedNoDataCellType => (Some(ct.noDataValue), Short.MaxValue)
-        case ct: UShortUserDefinedNoDataCellType => (Some(ct.widenedNoData.asInt), 65535)
-        case ct: IntUserDefinedNoDataCellType => (Some(ct.noDataValue), Int.MaxValue)
-        case ct: FloatUserDefinedNoDataCellType => (Some(ct.noDataValue), Float.MaxValue)
-        case ct: DoubleUserDefinedNoDataCellType => (Some(ct.noDataValue), Double.MaxValue)
+        case BitCellType => (None, 1,0)
+        case ByteCellType => (None, Byte.MaxValue, Byte.MinValue)
+        case UByteCellType => (None, 255, 0)
+        case ShortCellType => (None, Short.MaxValue, Short.MinValue)
+        case UShortCellType => (None, 65535, 0)
+        case IntCellType => (None, Int.MaxValue, Int.MinValue)
+        case FloatCellType => (None, Float.MaxValue, Float.MinValue)
+        case DoubleCellType => (None, Double.MaxValue, Double.MinValue)
+        case ByteConstantNoDataCellType => (Some(byteNODATA), Byte.MaxValue, Byte.MinValue)
+        case UByteConstantNoDataCellType => (Some(ubyteNODATA), 255, 0)
+        case ShortConstantNoDataCellType => (Some(shortNODATA), Short.MaxValue, Short.MinValue)
+        case UShortConstantNoDataCellType => (Some(ushortNODATA), 65535, 0)
+        case IntConstantNoDataCellType => (Some(NODATA), Int.MaxValue, Int.MinValue)
+        case FloatConstantNoDataCellType => (Some(floatNODATA), Float.MaxValue, Float.MinValue)
+        case DoubleConstantNoDataCellType => (Some(doubleNODATA), Double.MaxValue, Double.MinValue)
+        case ct: ByteUserDefinedNoDataCellType => (Some(ct.noDataValue), Byte.MaxValue, Byte.MinValue)
+        case ct: UByteUserDefinedNoDataCellType => (Some(ct.widenedNoData.asInt), 255, 0)
+        case ct: ShortUserDefinedNoDataCellType => (Some(ct.noDataValue), Short.MaxValue, Short.MinValue)
+        case ct: UShortUserDefinedNoDataCellType => (Some(ct.widenedNoData.asInt), 65535, 0)
+        case ct: IntUserDefinedNoDataCellType => (Some(ct.noDataValue), Int.MaxValue, Int.MinValue)
+        case ct: FloatUserDefinedNoDataCellType => (Some(ct.noDataValue), Float.MaxValue, Float.MinValue)
+        case ct: DoubleUserDefinedNoDataCellType => (Some(ct.noDataValue), Double.MaxValue, Double.MinValue)
       }
     }
 
@@ -93,20 +93,23 @@ object GeneralUtils {
         case _: UByteCells => ShortUserDefinedNoDataCellType(Short.MaxValue)
         case _: ShortCells => IntUserDefinedNoDataCellType(Int.MaxValue)
         case _: UShortCells => IntUserDefinedNoDataCellType(Int.MaxValue)
-        case _: IntCells => DoubleConstantNoDataCellType
+        case _: IntCells => FloatConstantNoDataCellType
         case _: FloatCells => DoubleConstantNoDataCellType
         case _: DoubleCells => DoubleConstantNoDataCellType
       }
     }
 
     val dataType = cellTypeUnion(leftCellType,rightCellType)
-    val (maybeNodataLeft,maxLeft) = getNodataAndMax(leftCellType)
-    val (maybeNodataRight,maxRight) = getNodataAndMax(rightCellType)
+    val (maybeNodataLeft, maxLeft, minLeft) = getNodataMaxMin(leftCellType)
+    val (maybeNodataRight, maxRight, minRight) = getNodataMaxMin(rightCellType)
 
     if (maybeNodataLeft.isEmpty || maybeNodataRight.isEmpty){
       if (maybeNodataLeft.isDefined) {
         val nodataLeft = maybeNodataLeft.get
         if (maxRight < nodataLeft) {
+          dataType.withNoData(Some(nodataLeft))
+        }
+        else if (minRight > nodataLeft){
           dataType.withNoData(Some(nodataLeft))
         }
         else if (nodataLeft.isNaN){
@@ -124,6 +127,9 @@ object GeneralUtils {
       } else if (maybeNodataRight.isDefined) {
         val nodataRight = maybeNodataRight.get
         if (maxLeft < nodataRight) {
+          dataType.withNoData(Some(nodataRight))
+        }
+        else if (minLeft > nodataRight){
           dataType.withNoData(Some(nodataRight))
         }
         else if (nodataRight.isNaN){
@@ -159,6 +165,9 @@ object GeneralUtils {
         } else if (nodataLeft > nodataRight) {
           if (maxRight < nodataLeft) {
             dataType.withNoData(Some(nodataLeft))
+          }
+          else if (minLeft > nodataRight ){
+            dataType.withNoData(Some(nodataRight))
           } else {
             if (isUnSigned(leftCellType) || isUnSigned(rightCellType)) {
               if (leftCellType.bits >= rightCellType.bits){
@@ -173,6 +182,9 @@ object GeneralUtils {
         } else { // nodataRight > nodataLeft
           if (maxLeft < nodataRight) {
             dataType.withNoData(Some(nodataRight))
+          }
+          else if (minRight > nodataLeft ){
+            dataType.withNoData(Some(nodataLeft))
           } else {
             if (isUnSigned(leftCellType) || isUnSigned(rightCellType)) {
               if (leftCellType.bits >= rightCellType.bits){