+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Experimental support for improved CORSA compression and decompression ([#702](https://github.com/Open-EO/openeo-geotrellis-extensions/issues/702))
 - Support BigTiff output ([#792](https://github.com/Open-EO/openeo-geotrellis-extensions/issues/792))
 - Support treating missing angle bands file as soft error ([#692](https://github.com/Open-EO/openeo-geotrellis-extensions/issues/692))
+- Fail fast when produced GeoTiff is corrupt ([#794](https://github.com/Open-EO/openeo-geotrellis-extensions/issues/794))
 
 ### Changed
 
@@ -34,6 +35,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `save_result` : add metadata for NetCDF assets ([#406](https://github.com/Open-EO/openeo-geotrellis-extensions/issues/406))
 - `regrid`/`apply_neighborhood`: optimize target partitioner based on target tile size, reducing executor memory pressure for large tiles ([#626](https://github.com/Open-EO/openeo-geotrellis-extensions/issues/626))
 - Improve default values for `DataCubeParameters` ([#748](https://github.com/Open-EO/openeo-geotrellis-extensions/issues/748))
+- use dataType and nodata from STAC metadata and take a union of the dataTypes for the datacube ([#658](https://github.com/Open-EO/openeo-geotrellis-extensions/issues/658))
 
 ### Removed